sketch the theoretical results section
Wrote the introduction. The meat is not there yet.

The N = 4 case (relying on the PhD thesis of the guy whose name I still
can't recall) is Robin's field of expertise.

The N = 5 case, dealing with the polynomials and curves, is my field of
expertise, and will be populated soon.