rewrote the general results subsection (RED ALERT)
@RobinZ16

Almost everything is completely rewritten to ensure styles are
consistent and important steps are not skipped (the usage of each
condition should be stated clearly and unambiguously, eliminating any
handwavy argument --- three trivialities unchecked consists a no-pass).

I stopped checking (and hence rewriting) at the sentence

"If this is true in general, then it follows that all periodic points
for \phi_c (with c rational) in Galois extensions to Q fields with
degree dividing by their period correspond to rational points in C_0(N),
since the first case produces rational trace t = sum_{i=1}^N z_i."

I thought for about five (or ten) minutes without success --- maybe I'm
just too tired after sitting and writing for the past twelve hours, but
I don't want to spend the rest of the night on this, especially because
the computational section is not yet expanded and rewritten.

(In fact, I'm under the impression that (ii) does not hold might
automatically guarantee the trace is rational — this is the case in N =
4, where you don't really need x1 = bar(x3). As long as you have x1 =
bar(something), the trace will be rational. Just a thought.)

I won't be able to proceed without a reply on this matter.