numerous changes, mostly not math
The only substantial math change is regarding the number of rational
points on P(a, b) = 0. I checked the rational points in detail and found
that we know five regular points and three singular points (all of which
are nodes):

Regular: (0, 0), (4, 1/3), (1, 8/3), (6, 9), [0, 1, 0] (infty);
Nodes: (3, 0), [1, 3, 0] (infty), [1, 0, 0] (infty);

So we expect at most 11 rational points on the normalization.